+++ src/performance/scala/org/economicsl/agora/markets/auctions/mutable/orderbooks/PriorityOrderBookMicroBenchmark.scala
@@ -17,7 +17,7 @@ package org.economicsl.agora.markets.auctions.orderbooks.mutable
 
 import org.economicsl.agora.markets.tradables.orders.ask.{AskOrder, LimitAskOrder}
 import org.economicsl.agora.OrderGenerator
-
+import org.economicsl.agora.markets.auctions.mutable.orderbooks.PriorityOrderBook
 import org.scalameter.api._
 import org.scalameter.{Bench, Gen}
