+++ src/test/scala/org/economicsl/agora/markets/auctions/concurrent/orderbooks/SortedOrderBookSpec.scala
@@ -13,11 +13,10 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package org.economicsl.agora.markets.auctions.orderbooks.concurrent
+package org.economicsl.agora.markets.auctions.concurrent.orderbooks
 
-import org.economicsl.agora.markets.auctions.orderbooks
-import org.economicsl.agora.markets.tradables.orders.bid.{BidOrder, LimitBidOrder, MarketBidOrder}
 import org.economicsl.agora.markets.tradables.Tradable
+import org.economicsl.agora.markets.tradables.orders.bid.{BidOrder, LimitBidOrder, MarketBidOrder}
 
 
 class SortedOrderBookSpec extends org.economicsl.agora.markets.auctions.orderbooks.OrderBookSpec[BidOrder, SortedOrderBook[BidOrder]] {
