+++ src/test/scala/org/economicsl/agora/markets/auctions/concurrent/orderbooks/parallel/OrderBookSpec.scala
@@ -13,13 +13,10 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package org.economicsl.agora.markets.auctions.orderbooks.parallel.concurrent
+package org.economicsl.agora.markets.auctions.concurrent.orderbooks.parallel
 
-import org.economicsl.agora.markets.auctions.orderbooks
-import org.economicsl.agora.markets.tradables.orders.ask.{AskOrder, LimitAskOrder, MarketAskOrder}
 import org.economicsl.agora.markets.tradables.Tradable
-
-import scala.util.Random
+import org.economicsl.agora.markets.tradables.orders.ask.{AskOrder, LimitAskOrder, MarketAskOrder}
 
 
 class OrderBookSpec extends org.economicsl.agora.markets.auctions.orderbooks.OrderBookSpec[AskOrder, OrderBook[AskOrder]] {
