+++ src/test/scala/org/economicsl/agora/markets/auctions/mutable/continuous/TestPostedPriceAuction.scala
@@ -13,10 +13,10 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package org.economicsl.agora.markets.auctions.continuous
+package org.economicsl.agora.markets.auctions.mutable.continuous
 
 
-import org.economicsl.agora.markets.auctions.{TestBuyerPostedPriceAuction, TestSellerPostedPriceAuction}
+import org.economicsl.agora.markets.auctions.TestSellerPostedPriceAuction
 import org.economicsl.agora.markets.tradables.orders.ask.LimitAskOrder
 import org.economicsl.agora.markets.tradables.orders.bid.LimitBidOrder
 import org.economicsl.agora.markets.tradables.Tradable
