+++ src/test/scala/org/economicsl/agora/markets/auctions/mutable/orderbooks/PriorityOrderBookSpec.scala
@@ -13,28 +13,42 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package org.economicsl.agora.markets.auctions.orderbooks.concurrent
+package org.economicsl.agora.markets.auctions.mutable.orderbooks
 
+import java.util.UUID
+
+import org.economicsl.agora.markets.auctions.mutable.orderbooks.PriorityOrderBook
 import org.economicsl.agora.markets.auctions.orderbooks
 import org.economicsl.agora.markets.tradables.orders.ask.{AskOrder, LimitAskOrder, MarketAskOrder}
 import org.economicsl.agora.markets.tradables.Tradable
 
+import scala.collection.mutable
 
-class OrderBookSpec extends org.economicsl.agora.markets.auctions.orderbooks.OrderBookSpec[AskOrder, OrderBook[AskOrder]] {
 
-  def orderBookFactory(tradable: Tradable): OrderBook[AskOrder] = OrderBook[AskOrder](tradable)
+class PriorityOrderBookSpec extends orderbooks.OrderBookSpec[AskOrder, PriorityOrderBook[AskOrder, mutable.Map[UUID, AskOrder]]] {
 
-  feature(s"A concurrent.OrderBook should be able to add ask orders.") {
+  def orderBookFactory(tradable: Tradable) = PriorityOrderBook[AskOrder](tradable)
+
+  feature("A PriorityOrderBook should be able to be built from specified type parameters.") {
+
+    scenario("Creating an OrderBook using generic constructor.") {
+      val orderBook = PriorityOrderBook[AskOrder, mutable.WeakHashMap[UUID, AskOrder]](validTradable)
+      assert(orderBook.isInstanceOf[PriorityOrderBook[AskOrder, mutable.WeakHashMap[UUID, AskOrder]]])
+    }
+  }
+
+  feature(s"A mutable.PriorityOrderBook should be able to add ask orders.") {
 
     val orderBook = orderBookFactory(validTradable)
 
-    scenario(s"Adding a valid ask order to an concurrent.OrderBook.") {
+    scenario(s"Adding a valid ask order to a mutable.PriorityOrderBook.") {
       val order = orderGenerator.nextLimitAskOrder(None, validTradable)
       orderBook.add(order)
       orderBook.headOption should be(Some(order))
+      orderBook.existingOrders.headOption should be(Some((order.uuid, order)))
     }
 
-    scenario(s"Adding an invalid ask order to an concurrent.OrderBook.") {
+    scenario(s"Adding an invalid ask order to a mutable.PriorityOrderBook.") {
       val invalidOrder = orderGenerator.nextLimitAskOrder(None, invalidTradable)
       intercept[IllegalArgumentException] {
         orderBook.add(invalidOrder)
@@ -43,30 +57,32 @@ class OrderBookSpec extends org.economicsl.agora.markets.auctions.orderbooks.Ord
 
   }
 
-  feature(s"A concurrent.OrderBook should be able to remove ask orders.") {
+  feature(s"A mutable.PriorityOrderBook should be able to remove an AskOrder.") {
 
-    scenario(s"Removing an existing ask order from an concurrent.OrderBook.") {
+    scenario(s"Removing an existing ask order from a mutable.PriorityOrderBook.") {
       val order = orderGenerator.nextLimitAskOrder(None, validTradable)
       val orderBook = orderBookFactory(validTradable)
       orderBook.add(order)
       val removedOrder = orderBook.remove(order.uuid)
       removedOrder should be(Some(order))
       orderBook.headOption should be(None)
+      orderBook.existingOrders.headOption should be(None)
     }
 
-    scenario(s"Removing an ask order from an empty concurrent.OrderBook.") {
+    scenario(s"Removing an ask order from an empty mutable.PriorityOrderBook.") {
       val order = orderGenerator.nextLimitAskOrder(None, validTradable)
       val orderBook = orderBookFactory(validTradable)
       val removedOrder = orderBook.remove(order.uuid)  // note that order has not been added!
       removedOrder should be(None)
       orderBook.headOption should be(None)
+      orderBook.existingOrders.headOption should be(None)
     }
 
   }
 
-  feature(s"A concurrent.OrderBook should be able to find an AskOrder.") {
+  feature(s"A mutable.PriorityOrderBook should be able to find an AskOrder.") {
 
-    scenario(s"Finding an existing LimitAskOrder in an concurrent.OrderBook.") {
+    scenario(s"Finding an existing LimitAskOrder in an mutable.PriorityOrderBook.") {
       val limitOrder = orderGenerator.nextLimitAskOrder(None, validTradable)
       val marketOrder = orderGenerator.nextMarketAskOrder(None, validTradable)
       val orderBook = orderBookFactory(validTradable)
@@ -76,7 +92,7 @@ class OrderBookSpec extends org.economicsl.agora.markets.auctions.orderbooks.Ord
       foundOrder should be(Some(limitOrder))
     }
 
-    scenario(s"Finding a MarketAskOrder in an concurrent.OrderBook containing only LimitAskOrder instances.") {
+    scenario(s"Finding a MarketAskOrder in an mutable.PriorityOrderBook containing only LimitAskOrder instances.") {
       val limitOrder = orderGenerator.nextLimitAskOrder(None, validTradable)
       val anotherLimitOrder = orderGenerator.nextLimitAskOrder(None, validTradable)
       val orderBook = orderBookFactory(validTradable)
@@ -88,29 +104,31 @@ class OrderBookSpec extends org.economicsl.agora.markets.auctions.orderbooks.Ord
 
   }
 
-  feature(s"A concurrent.OrderBook should be able to remove the head AskOrder.") {
+  feature(s"A mutable.PriorityOrderBook should be able to remove the head AskOrder.") {
 
-    scenario(s"Removing the head AskOrder from an concurrent.OrderBook.") {
+    scenario(s"Removing the head AskOrder from a mutable.PriorityOrderBook.") {
       val order = orderGenerator.nextLimitAskOrder(None, validTradable)
       val orderBook = orderBookFactory(validTradable)
       orderBook.add(order)
       val removedOrder = orderBook.remove()
       removedOrder should be(Some(order))
       orderBook.headOption should be(None)
+      orderBook.existingOrders.headOption should be(None)
     }
 
-    scenario(s"Removing the head AskOrder from an empty concurrent.OrderBook.") {
+    scenario(s"Removing the head AskOrder from an empty mutable.PriorityOrderBook.") {
       val orderBook = orderBookFactory(validTradable)
       val removedOrder = orderBook.remove()  // note that order has not been added!
       removedOrder should be(None)
       orderBook.headOption should be(None)
+      orderBook.existingOrders.headOption should be(None)
     }
 
   }
 
-  feature(s"A concurrent.OrderBook should be able to filter its existingOrders.") {
+  feature(s"A mutable.PriorityOrderBook should be able to filter its existingOrders.") {
 
-    scenario(s"Finding all existing MarketAskOrder instances an concurrent.OrderBook.") {
+    scenario(s"Finding all existing MarketAskOrder instances an mutable.PriorityOrderBook.") {
       val limitOrder = orderGenerator.nextLimitAskOrder(None, validTradable)
       val marketOrder = orderGenerator.nextMarketAskOrder(None, validTradable)
       val orderBook = orderBookFactory(validTradable)
@@ -120,7 +138,7 @@ class OrderBookSpec extends org.economicsl.agora.markets.auctions.orderbooks.Ord
       filteredOrders should be(Some(Iterable(marketOrder)))
     }
 
-    scenario(s"Finding all MarketAskOrder in an concurrent.OrderBook containing only LimitAskOrder instances.") {
+    scenario(s"Finding all MarketAskOrder in an mutable.PriorityOrderBook containing only LimitAskOrder instances.") {
       val limitOrder = orderGenerator.nextLimitAskOrder(None, validTradable)
       val anotherLimitOrder = orderGenerator.nextLimitAskOrder(None, validTradable)
       val orderBook = orderBookFactory(validTradable)
@@ -131,5 +149,6 @@ class OrderBookSpec extends org.economicsl.agora.markets.auctions.orderbooks.Ord
     }
 
   }
-  
+
 }
+
